Refactor DataLoader to remove label column handling, streamline data loading, and integrate Cleaner for data preprocessing; enhance RecordClassifier evaluation method for improved accuracy reporting.
